+++ relay-server/src/services/buffer/stack_provider/mod.rs
@@ -58,6 +58,10 @@ pub trait StackProvider: std::fmt::Debug {
     /// Returns the total count of the store used by this [`StackProvider`].
     fn store_total_count(&self) -> impl Future<Output = u64>;
 
+    /// Returns the number of bytes the storage occupies. Will return `None` if no
+    /// reliable information can be provided.
+    fn total_size(&self) -> Option<u64>;
+
     /// Returns the string representation of the stack type offered by this [`StackProvider`].
     fn stack_type<'a>(&self) -> &'a str;